requirements.txt
python-gitlab==1.14.0
Jinja2==2.10.3
Click==7.0
